Move conversions to Expr
